Add the access to the MTU value in TCP sockets (linux-specific).
The MTU is measured by the kernel and read via getsockopt(). Then
inet:getopts/2 accepts the option 'mtu' in argument and return the
mtu value.
Event though the kernel does the Right Thing (TM) on its own, it
is usefull to retrieve the MTU for other purposes (eg
complementary UDP socket).